% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetDecoyQC-package.R
\docType{package}
\name{TargetDecoyQC-package}
\alias{TargetDecoyQC}
\alias{TargetDecoyQC-package}
\title{TargetDecoyQC: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Milan Malfait \email{milan.malfait94@gmail.com} (\href{https://orcid.org/0000-0001-9144-3701}{ORCID})

}
\keyword{internal}
